Home work class work active deactive, dashboard, drawer profile tab and attendance shortage ui done

+++ lib/staffs/homework_classwork/model/homework_attachment_model.dart
@@ -32,7 +32,7 @@ class HomeWorkAttachmentModelValues {
   int? ivrmrTId;
   int? asmaYId;
   int? amsTId;
-  int? marks;
+  double? marks;
   int? ihWAssignmentNo;
   int? ismSId;
   int? ivrmuLId;
